+++ python/ray/tests/test_experimental_client_references.py
@@ -1,4 +1,4 @@
-from ray._private.ray_client_helpers import ray_start_client_server
+from ray.experimental.client.ray_client_helpers import ray_start_client_server
 from ray.test_utils import wait_for_condition
 import ray as real_ray
 from ray.core.generated.gcs_pb2 import ActorTableData
